p4/python/requirements.txt
scipy
numpy
matplotlib
pyserial
